treewide: resolve compiler warnings
ECGenericObjectTable.cpp: In member function 'ECRESULT ECGenericObjectTable::UpdateCategoryMinMax(sObjectTableKey&, ECCategory*, unsigned int, propVal*, unsigned int, bool*)':
ECGenericObjectTable.cpp:2832:31: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  if(lpsSortOrderArray->__size <= i || !ISMINMAX(lpsSortOrderArray->__ptr[i].ulOrder))
ECICSHelpers.cpp: In member function 'ECRESULT ECGetContentChangesHelper::MatchRestrictions(const std::vector<char**>&, const std::vector<long unsigned int*>&, restrictTable*, std::set<SOURCEKEY>*)':
ECICSHelpers.cpp:1031:79: warning: format '%d' expects argument of type 'int', but argument 3 has type 'std::vector<char**>::size_type {aka long unsigned int}' [-Wformat=]
  ec_log(EC_LOGLEVEL_ICS, "MatchRestrictions: matching %d rows", db_rows.size());
ECSecurity.cpp: In member function 'virtual ECRESULT ECSecurity::SetRights(unsigned int, rightsArray*)':
ECSecurity.cpp:815:15: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  if (ulErrors == lpsRightsArray->__size)
ECNotifyMaster.cpp: In static member function 'static void* ECNotifyMaster::NotifyWatch(void*)':
ECNotifyMaster.cpp:371:10: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   if (hr == KCWARN_CALL_KEEPALIVE) {
epclient.cpp: In function 'HRESULT MSGServiceEntry(HINSTANCE, LPMALLOC, LPMAPISUP, ULONG, ULONG, ULONG, ULONG, LPSPropValue, LPPROVIDERADMIN, MAPIERROR**)':
epclient.cpp:564:9: warning: variable 'bShowAllSettingsPages' set but not used [-Wunused-but-set-variable]
WSMAPIPropStorage.cpp: In member function 'virtual HRESULT WSMAPIPropStorage::HrUpdateSoapObject(MAPIOBJECT*, saveObject*, convert_context*)':
WSMAPIPropStorage.cpp:320:37: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    if(lpsSaveObj->modProps.__size+1 > lpsMapiObject->lstModified->size() ) {
WSMAPIPropStorage.cpp: In member function 'virtual ECRESULT WSMAPIPropStorage::ECSoapObjectToMapiObject(saveObject*, MAPIOBJECT*)':
WSMAPIPropStorage.cpp:518:11: warning: variable 'ec' set but not used [-Wunused-but-set-variable]
  ECRESULT ec;
vtimezone.cpp: In function 'HRESULT HrZoneToStruct(icalcomponent_kind, icalcomponent*, TIMEZONE_STRUCT*)':
vtimezone.cpp:146:50: warning: variable 'rDate' set but not used [-Wunused-but-set-variable]
  icalproperty *tzFrom, *tzTo, *rRule, *dtStart, *rDate;

+++ provider/client/ECNotifyMaster.cpp
@@ -368,7 +368,7 @@ void* ECNotifyMaster::NotifyWatch(void *pTmpNotifyMaster)
 		notificationArray *pNotifyArray = NULL;
 
 		hr = pNotifyMaster->m_lpTransport->HrGetNotify(&pNotifyArray);
-		if (hr == KCWARN_CALL_KEEPALIVE) {
+		if (static_cast<unsigned int>(hr) == KCWARN_CALL_KEEPALIVE) {
 			if (bReconnect) {
 				TRACE_NOTIFY(TRACE_WARNING, "NotifyWatch::Reconnection", "OK connection: %d", pNotifyMaster->m_ulConnection);
 				bReconnect = false;
